Lots of cleanup for the CDR logging, and improves the permission request process
There was also a bug where there permissions dialog would get stuck in a loop because the push_notification permission always returned true for should show rationale, but never showed the permission request dialog. I added a hack fix for this, but only for the main dialog, not for the CDR dialog.